Add support for commands not written in Haskell
Some people have expressed that they would rather program in another
language than Haskell. We have therefore implemented support for
programming in other languages.

The plugins should report themselves in the .json file, they will then be added
to the list of functions on Dikunt's start.

When a plugin is running the bot has set the environment variable MESSAGE to the
value of the message the bot is processing.