General changes to plugin management
A directory name can now be given to Dikunt and the bot will look for
executables there. That means that users no longer has to extend their
path to run the bot.

Plugins can be turned on and off on the command line and not only in the
configuration of the bot.

Rewrite of Help plugin. It will now only give help to the plugins
running and not all plugins in the configuration file.